ONNX caused issues, switched to joblib

+++ variables.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
 class Variables(BaseModel):
-    hour: int
-    day: int
-    month: int
+    hour: float
+    day: float
+    month: float
